PDF interpreter - fix 32-bit integer overflow detection
Bug #707288 ""Error: Password did not work" with apparently valid pdf"

The PDF file has a Permissions value written as an unsigned integer
instead of a signed integer. Because the high bits of the value are
set to 1 this results in a very large number (dhould be a small
negative number).

We were setecting overflow of an unsigned 32-bit integer one digit too
early, leading to us replacing the value with the maximum integer.
Because this value is used as part of the key generation this caused
the decryption to fail.