XPS interpreter - assume resolution of 0 to mean 96
Bug 707290 " image in xps convert to pdf failed."

The JPEG image in the XPS file has XDensity = YDensity = 0, which
causes us to skip the image.

MuPDF assumes in this case that the resolution should be 96, and we
adopt that here. MuPDF also reads the APP13 and EXIF markers to find
other potential resolutions, and apparently prefers EXIF to the JPEG
data  anyway. This commit does not attempt that and the bug is altered
to an enhancement to reflect this.