pdfwrite device - prevent buffer underrun with corrupt TT font
OSS-fuzz #63648

A corrupted TrueType font could execute a jump instruction where the
distance to jump moved the instruction pointer back before the start
of the glyph program.

Detect that here, and if it happens move the instruction pointer to the
end of the glyph program so we terminate processing it. We can't return
an error at this level.