rinkj & xcf devices - prevent buffer overflow
Bug #707221 "RINKJ device: buffer overflow with "SeparationColorNames" parameter"
Bug #707222 "XCF device: buffer overflow with "SeparationColorNames" parameter"

These are the same problem in two different devices, so I've chosen
to merge them into a single commit.

The basics of this commit were supplied by Zdenek Hutyra and check
the number of spots and the number of existing named separations to
ensure we don't exceed the maximum number of separations supported by
the device.

In addition to Zdenek's original patch I have added code to create new
copies of the separation strings containing the names, rather than
simply using the supplied names as the old code did. This is because
those strings are not guaranteed to persist, and we must therefore
not retain a reference to them.

These new strings are stored in non-gc memory and freed when the device
is closed, or they are replaced.

squash again