oss-fuzz 63650: Tweak decoding of hex strings
Previously, pdfi would onlly allow hex strings in objects of Postscript syntax
(Type 1 fonts and CMaps) with an even number of digits, but this could cause
a problem if a CMap had an (obviously invalid) hex string of one character. We
could end up with a zero length key in the CMap, causing problems with the later
string decoding, resulting in an infinite loop.

This changes it so if we encounter an odd number of characters in a hex string,
we'll try to treat the final character as a nibble.

For safety, also add a check to ensure we cannot add a zero length key string
to a CMap.