pdfwrite - fix bug in /DeviceN colour conversion
No bug report for this, the problem was reported by a customer
running LinuxONE on an IBM s390. The file caused a seg fault on that
system, but was not reproducible on any other setup, including other
big-endian systems.

Folloing debugging work by the customer, and inspection of the relevanr
code, it looked likely that the problem was due to a copy/paste error
when originally writing the colour conversion code. The DeviceN case
has been copied form the /Separation case, and special handling for
/All and /None left in place.

We don't need that special handling for DeviceN, and worse than that,
the code was modifying the 'params' union using the separation member
when in fact it is the DeviceN member which should be altered. This, in
effect, corrupts the contents.

Removing the special case fixes the problem for the customer and looks
like the correct thing to do anyway. This commit does that and changes
the 'sep_space' variable to 'devicen_space' to make it less confusing.