Add some utility functions to allow devices to be de-cached
See bug 707309 "Can't reset page counter for 'png16m' device"

The problem is the behaviour of setpagedevice (and, historically,
finddevice). For some unknown reason but presumably an opitmisation
when we first create an instance of a device we cache it in a dictionary
called 'devicedict'. If we restore the device away, that instance still
persists, and the next time we look up a device we start by looking
in devicedict. Since there is an entry, we use it.

The problem is that this is not what a user should expect to happen
when a device is restored away. In general this wouldn't be a problem
but any parameters of the device which are not default will retain
their modified values.

Previously a (ugly) way around this was to use 'findprotodevice
copydevice' instead of simply restoring the device away, in order to get
a new fresh instance, but we took findprotodevice away.

This commit contains two utility functions which remove a named device
instance from the device cache, so that the next time we use
'<</OutputDevice /Name >>setpagedevice' we will get a fresh new device,
and a function to remove *all* instances except the current device from
the cache.

In future I anticipate removing this caching feature altogether, unless
someone can explain what its function is, because it complicates
matters and the current behaviour is at best unexpected and at worst
plain wrong.