PDF interpreter - detect and avoid a circular reference
OSS-fuzz #63874

The fuzzed file has a Link annotation whic correctly points to a URI
as an indirect object. However, the file also contains an object with
the same object and generation as the indirect URI, and because that
is later in the file, that's the one we use after repairing it.

That object is the Root dictionary, which contains the Pages array, and
the Pages contain a reference back to the Pages array, making this a
circular reference. Which we didn't detect.

This commit amends the loop detection to account for this case. This
causes differences in a few files with pdfwrite, these files threw a
circular reference before this commit but with this commit the error
occurs in a slightly different place leading to a slightly different
output file. Not any more wrong, just different.