pdfwrite - Implement XRefStm and ObjStm support
I thought there was an enhancement bug report for this, but I'm unable
to find it, possibly it was closed in the past.

This commit adds the ability for pdfwrite to use an XRef stream when
writing the output, and also to store non-stream objects in an ObjStm.

Together these can save up to around 25% on the size of an output file,
though there is some cost in performance both writing the output file
and when reading it with a PDF consumer.

There are checks to ensure that we don't use these features when writing
too low a level of PDF file (<1.5), or when Linearizing (optimize for
fast web view) as we don't support ObjStms with linearization. We also
disable ObjStms when the output is ps2write, as these are incompatible.
(We don't write an xref dor ps2write, so it doesn't matter whether we
disable XRef streams or not).

Also document the two new controls. It is possible that these features
will introduce bugs, so they can be disabled.