Graphics library - prevent integer overflow in page buffer calculations
The calculation in gdev_prn_allocate() deciding whether we have
sufficient memory to run the entire page in memory, or need
to revert to a clist, could overflow because the variable was
defined as a ulong, but the inputs are size_t. On Windows a
ulong is 32-bits but a size_t (on a 64-bit build) is 64-bits.

Promote the intermediate variable to a size_t to match the inputs
and add a check to make sure that we don't overflow it anyway.

While it's unlikely we'd overflow a 64-bit variable, on a 32-bit
build it's quite conceivable that we could.

Thanks to Peter Skarpetis for reporting the problem, tracking down
the fault and supplying the fix.