PDF interpreter - fix optional content groups with marked content
Bug #707325 "New engine renders extraneous elements"

The problem here is that the PDF Reference overloads the 'EMC' operator,
It uses BDC for 'property lists' (eg optional content) and BMC for
marked content, but EMC is used to terminate both.

Because we ignore marked content when rendering we did not increment
the 'BMClevel' on encountering a BMC operator, but we do decrement it
when we get an EMC. The problem is, if we've incremented it because
of a BDC, then get a BMC/EMC pair, the EMC of that pair will decrement
the level causing us to terminate the optional content group.

Fix that here by simply incrememting BMClevel when we get a 'BMC'.