PDF interpreter - detect Patterns using blend modes other than Normal/Compatible
Apparently transparent Patterns which use Normal/Compatible blend
modes and then switch to a different blend mode during the course of
the Pattern can't work with a bitmap pattern tile, because we always
apply the blend mode to the entire tile.

So we need to detect patterns which use these blend modes and flag them,
because we can't detect whether the entire pattern is in the non-Normal
mode or it switches part way, we will assume that any pattern which uses
a non-Normal blend mode needs to be treated this way.

When we are "checking" a resource dictionary we can no longer
exit as soon as we know that we have transparency. Previously,
we used to be checking just for transparency usage, but now we're
looking for non-normal blend modes too.

As such, we have to keep working after we find transparency.