pdfwrite - turn off ObjStms and XRef Streams silently for PDF/X and PDF/A
If we're producing PDF/X, which is limited to PDF 1.3, or PDF/A-1
(PDF 1.4), then we can't use ObjStms and XRef streams. The code did
already cater for this, but it emitted a warning that they were
incompatible, which might be confusing, so turn them off before we get
to that point.