pdfwrite - don't check ICC components
This was causing an error with -sOutputICCProfile, because (for reasons
I don't fathom) we get the OutputICCProfile, and process it in the
default handler, before we get a second call to put_params() with the
ColorConversionStrategy.

Because we set the ProcessColorModel when we set the Color Conversion,
this meant the first call would throw an error. There isn't a good
solution to this as far as I can see, so simply don't check the
components, we'll haev to rely on the user getting it right.