pdfwrite & PDF interpreter - fix Group colour space
There's no bug report for this, and no specimen file, as the file came
from a customer with a request to treat it as confidential.

When run through pdfwrite, the resulting PDF file causes Adobe Acrobat
(all versions tested) to crash and burn. This turns out to be because
the PDF file has a transparency Group with a ColorSpace (CS) of
/Pattern, which is illegal.

This is due to commit 8c432810d36e24613b84f8506ab3929f9d13e612 which
builds on an earlier commit ea6290b302598f13e7fb4c29aff73657989e693d.
The problem occurs because the transparency code now sets the
group_color_type member of the params structure to be either ICC or
GRAY_SCALE when creating a mask. But pdfwrite relies upon that member
being 'UNKNOWN' when the incoming PDF file has a group with no colour
space.

The pdfwrite device then uses the current colour space on the assumption
that it has been set by the interpreter to the Group colour space. But
in this case the original Group had no colour space, so we did not set
it, resulting in the code using whatever the current color space happens
to be, and in this case, that's a Pattern space.

I cannot change the transparency code, but it seems to me that this was
probably always the wrong test anyway. If the Group did not set a colour
space then we shoudl not have a colour space in the transparency params
and we should use that, rather than the group_color_type.

But that then trips over the PDF itnerpreter, where if we do not have a
colour space in the Group, we use the current colour space. That's
clealry wrong; if the Group has no space the transparency code should
use either the enclsoing object, or the device space.

So fix this by making the transparency params colour space memeber NULL
if the Group has no ColorSpace, and change the pdfwrite test to check
the ColorSpace member of the params structure rather than the group
color type. Also, *use* the ColorSpace member of the params structure
rather than relying on the interpreter to have set the current colour
space to it.

This causes 3 files to increase in size, but no rendering differences.
The increase in size appears (surprisingly) to be because Groups with
an ICC space were presviously ending up without a ColorSpace at all,
and now we properly preserve the ICC space.