pdfwrite - fix OutputICCProfile
bug #707350 "Incorrect colour from ColorConversionStrategy with OutputICCProfile"

The problem is that when we set ColorConversionStrategy we don't
consider whether the user has supplied an OutputICCProfile they want us
to use, but simply set the defaults based on the colour model. This
overwrites the profile we should be using with the standard default.

This commit just notes whether the user has set OutputICCProfile, if
they have then we rely on gdev_vectore_put_params() to properly
initialise the device's ICC profile strcutre and set the requested
profile as the default. We no longer override it.

It is up to the user to make sure the profile is appropriate for the
colour model; the ICC colour management will give an error if it is
not.