Update tiffsep & tiffsep1 devices and documentation
Bug #707357 " Device tiffsep is missing the separation names"

This is deliberate, not a bug. The tiffsep device was altered in
commit 2e66582c527646015731cd0f4a0fb2e2d1559887 so that it only emits
the separation names (and CMYK equivalents) if PrintSpotCMYK is true.

Strangely, at the same time the tiffsep1 device was modified so that it
behaves the way the tiffsep device *previously* operated.

This leads to the two devices ehaving differently and the tiffsep1
device printing the same information twice if PrintSpotCMYK is set.

I've chosen to remove the code added to tiffsep1 and retain the code
in print_cmyk_equivalent() because the old code used the filename
sanitising code on the spot colour name so that difficult characters
like '%', ':', '\' and '/' are turned into underscores. We don't need to
do that for just printing the spot name.

Updated the documentation for tiffsep and tiffsep1 to indicate that this
is now under control of a switch.