JBIG2 decoder - terminate on EOF
No bug report for this one. The specimen file comes from a custoemr and
the content looks sensitive.

The problem is that the PDF interpreter was doing an unbounded read of
data through the JBIG2 decoder. If we happened to exactly hit the end
of file, so no bytes were read (in_size = 0) but last is true, then we
simply returned. We should try to complete the page, just as the code
already does when we hit EOF with a t least 1 byte available.