PDF interpreter - apply SubFileDecode to stream to buffer conversion
No bug report for this one. The specimen file comes from a custoemr and
the content looks sensitive.

The actual problem is fixed in the previous commit in the JBIG2 decoder
but in fact we should not have been hitting that condition.

When converting a stream into a memory buffer we were not limiting the
amount of data we read to the length of the stream (or terminating when
we hit 'endstream'). This was causing us to read on and attempt to
decode more data, until we hit EOF.

This commit adds a SubFileDecode filter to the filter chain. If we
know how much data we want to read, and the stream length is valid, then
we limit reading the stream to the number of bytes in the stream.
Otherwise we stop when we hit 'endstream'.

This doesn't solve any problems, but it should make some small speed
improvements and prevents errors of the type in the JBIG2 decoder.