Bug 707365: Fix tiffsep problem with limited separations.
The tiffsep device can be told to only produce a limited number
of separation files (and to do them in different orders), using
/SeparationOrder and /SeparationColorNames.

Unfortunately, this hasn't worked properly since the downscaler
refactor about 7 years ago. The downscaler was always being set
up to only produce a single plane, but prior to the refactor
it was 'coincidentally' passing through the others too.

The fix, implemented here, is to ensure that we always set the
downscaler up to produce all the planes, even if we then only
use a few of them. This is arguably slower than we'd like, but
given no one has spotted this being broken for 7 years, I can't
believe there are that many people that will complain.

Also, to fix it differently would require a further rewrite
in how the downscaler is called to allow the selection of planes
to be passed through it.