pdfwrite startup code - better handling of error with missing device
Reported by customer 470. If Ghostscript is built without the pdfwrite
device then the interpreter fails to start, resulting in a
stackunderflow in pop error.

The code in gs_pdfwr.ps to see if the pdfwrite device is built in was
assuming that an error from finddevice would always leave 2 objects on
the stack, so it popped those before skippint to the end of the file.

Some of the recent changes for security problems have meant that this is
no longer the case. In fact it was never a reliable test anyway.

This commit pushes a mark before testing the device, and does a
clearto mark afterwards to preserve the stack.