pdfwrite - fix a reference count
Bug #707433 "Converting PS to PDF causes seg fault"

This bug was introduced in this commit
d07b9485b2bb3a70552aee9407fcff8fa52698df

Unfortunately Nancy didn't give any information on how to reproduce the
memory leak. The countdown causes us to count down a reference to a
colour space which we have not counted up, which (much later!) results
in it being freed when there is still a reference to it, and eventually
results in a seg fault at a still later point when we try to use the
(now invalid) reference.

The reporter claims this works with 9.55 but if so, it is entirely by
coincidence, as the change predates that release (and 9.55 crashes for
me).

With no way to investigate the original commit, I'm effectively
reverting it, as a memory leak is less serious than a seg fault.