pdfwrite - force use of alternate space when sampling tint transform
Bug #707458 "Bad CMYK color conversion"

The problem is caused by the fact that the colour space is a /Separation
/Black with an alternate space of /DeviceGray. We sample the tint
transform at each end, converting to the requested space, and then
write a function which interpolates between.

For reasons I don't intend to try and research, the CMS decides that
it doesn't need to use the alternate space to convert Black to CMYK and
so the values we get are incorrect.

Fixed here using the same approach as already used for the /All and
/None inks; force the use of the alternate space.