Bug 706987: Fix bad PSD interaction with NumRenderingThreads
When using the command-line given in the bug:

gs -sDEVICE=psdcmyktags16 -sOutputFile=".\test.psd"
-sICCOutputColors="Cyan, Magenta, Yellow, Black, White, Tags"
-dNumRenderingThreads=4 -r300 tiger.pdf

the program dies with an error while setting up the
rendering threads.

The process goes something like: GS starts up, creates a
device. The device is setup in its default state. The
device is then opened; this updates num_components in
a non-standard (device-specific) way.

Then we copy the device for each of the rendering threads.
Each of the devices has a copy of the original devices
(non-standardly setup) color_info. We 'get_params' the current
state from the existing device, and then 'put_params' it
into the new device.

As part of this process, put_params will reset
color_info.num_components (undoing the device specific
setup). We then check that 'Colors' is the same as
color_info.num_components, and this fails, causing the exit
with an error code.

The fix, implemented here, is to add a new dev_spec_op,
"gxdso_adjust_colors". We call this in "put_params" where
we recalculate the color_info.

This enables the device to redo any 'non-standard' manipulations
before we check things like "Colors" for consistency.

Accordingly, we have moved the non-standard setup from psd_prn_open
into psd_spec_op, and the same code gets called from psd_prn_open
and from put_params.