pdfwrite - Report when a PDF/A cannot be produced because of SMask
Bug #707450 "Conformation to PDF/A fails specification ISO 19005-3:2012, clause: 6.2.4.3, test number 2 in VeraPDF"

The problem is that the input PDF contains an image with an SMask, the
PDF 1.7 Reference states on page 555 that these *must* haev a DeviceGray
colour space.

But we can only use Device spaces in PDF/A if we have a relevant profile
in the OutputIntent, ie we are converting all colour to DeviceGray.

If we hit this incompatibility we cannot ignore the feature (the output
would be incorrect), so just abor the PDF/A condition and tell the user.