Fix crash caused by halftone creation
This seems to be a problem which was already addressed in
gx_device_halftone_release. It is possible for the default order of a
halftone to be used by one of the components of the halftone, if there
is no halftone for that component.

The release code checks the component corder to see if it is the same as
the default order before releasing it. The last action is to release
the default order. This prevents a double free.

The code in gx_gstate_dev_ht_install, however, was not considering this
possible case. If a screen was not provided for a given component then
it would 'move' the default order from the existing halftone and null
the default.

But if a component was already using the default order then this would
defeat the check in the release code, the release code would therefore
release the corder, thinking that it was not shared, while the new
halftone now held a pointer to it. Eventually we would try to release
this new halftone, and then would try to release the corder which we
had already released.....

NB it is the 'bit_data' member of the corder which is actually freed.

I think this 'move' is an optimisation, its cheaper to copy the pointers
than to make a new allocation and copy the contents. But since it fails
under these conditions I really don't think it is worth it. For this to
be succesful we'd need to test all the components before we copied the
default corder. We already have code to copy the corder so let's just
copy it unconditionally.