Bug 707485: Clean up exec stack on glyphshow error
The other "show" operations already cleanup the exec stack on error, in case
the error in question causes the operator to be retried (the most obvious of
these is stackoverflow error).

We do this in each operator, rather than folding it into op_show_continue_pop()
because the different operators push different numbers of objects into the
exec stack.