Fixes for saved-pages operation SEGVs with psd device.
Saved page operation relies on being able to print to clist
files, and play them back "at some point in the future" (rather than
"immediately" as usually happens).

clist files are both color and resolution dependent. If the device
can change its color model in the meantime (say, by adding spots),
then the clist files will be invalidated.

Since the change to fix PSD device operation with NumRenderingThreads
went in (about 16 commits ago?) the saved-pages-test overnight cluster
test has been showing SEGVs.

What was happening before, was that these tests would do the
saving of the pages just fine, but then when they came to play
them back, the color model for the page would not work, and so
we'd error out with a rangecheck.

The NumRenderingThreads fix involved a dev_spec_op being called from
within putparams to allow the device to tweak its color setup. The
PSD device is the only device, currently, to make use of it.

This means that when we putparams on a page reload, the PSD device
correctly resets its color setup, and we no longer range check.
Instead, we crash.

Why do we crash? Well, I've investigated this and found some things,
but not really a solution.

An example file that shows the problem:

  membin/gswin64c.exe -sOutputFile=out%d.psd --saved-pages-test
  -dMaxBitmap=400000000 -sDEVICE=psdcmyk -r72 -Z:
  -sDEFAULTPAPERSIZE=letter -dNOPAUSE -dBATCH -K2000000 -dClusterJob
  -dJOBSERVER ../perf-testing-gpdl/pdf/Ad_InDesign.pdf

Firstly, the work done by the PSD device on this callback relies
on the devn setup, so this needs to be correct before calling
putparams. Hence, in do_page_load, call gs_putdeviceparams
after we have set the devn separations, not before.

This seems right, doesn't hurt anything, but doesn't solve the
problem.

Secondly, when we restore the clist state to play back a saved
page, we are relying on the buffer device being already set up
correctly. This is sadly not the case for planar devices, as we
were not restoring the line pointer offset. This results in us
writing off the end of the buffer device.

We now store the line_ptrs_offset with the page, and restore that.
I am not 100% convinced that this is all we should be doing
here, but it's definitely wrong, and so this can't be worse.
It seems to solve the problem seen with the buffer device, but
we still crash later on.

Similarly, store and restore num_planar_planes with the page, as
this should count as part of the color setup.

With all these fixes in, it still does not work. With the example
above, it crashes in the clist reading when printing page 2.
It looks to me like the color setup for the device is correct. I
am at a loss to explain this, but having spent the best part of
a week looking into this, I need to move on.

Accordingly, I've used the gxdso_supports_saved_pages option
to disable saved pages with the PSD device. Previously files
would have been erroring out, so now they will just print
without saved-pages.

If someone ever actually needs to use saved-pages with the PSD
device in future, we can invest the time in trying to understand
and fix this properly then.