XPS interpreter - pass hyperlinks to the pdfwrite device
Bug #691342 "(Hyper)Links lost on conversion to PDF"

This should work for fill and stroke operations (paths basically). I
believe it is possible to have a hyperlink with text, but I don't have
an example for that.

This will only give useful results with external URLs, it won't work as
an internal link to the current document.