PDF interpreter - apply decryption when creating filtered memory streams
Bug #707516 " Issue with Ghostscript Version 10.02.1 - Visual Degradation in Merged PDFs"

The bug report claims that text goes missing when the specimen file is
run through GS to the pdfwrite device; I cannot reproduce any such
error.

However, the file is encrypted, and there are a number of shadings
which use function type 0 (sampled function). These store the sampled
data in streams, which are uncompressed.

Because there are no filters associated with the streams we end up
not applying decryption either, and so the data is incorrect.

Fix by applying decryption, even if there are no filters.