PDF interpreter - improved warning
Bug 707516; the attached file has a technically invalid Encrypt
dictionary. It has a /V of 4 and includes a /Length. The spec says that
/Length is optional, deprecated and only present when /v is 2 or 3.

We were using the same warning for an encryption dictionary with a /V
of 3 where /Length (which is permitted when V is 3) was not a valid
value (between 40 and 128 and a multiple of 8).

Define a new warning when the /Length value simply should not be
present at all and use it.