Ensure filename argument to sfopen is NULL-terminated
Bug 707511 "Fix sfopen() calls for unterminated string"

This is a different patch to that supplied in the bug report. This one
does not define a new function, does not use malloc and free, and also
avoids functions like strnlen (and reviewed all uses of sfopen and
located another suspect)

Instead, use a heap buffer for the filename, or use the supplied string
length where required to allocate buffers, rather than strlen.