Update tesseract/leptonica.
At this point both tesseract and leptonica should be on the
artifex-10.03.0 tags.

At the moment, we seem to be getting an 'Error: /undefined in --runpdf--'
after any PDF file is run to the ocr devices, but this seems to
happen regardless of whether I update tesseract/leptonica.

Accordingly, I'm going to commit the update, and we'll look at
fixing the error independently.