Bug 707538: Avoid stack buffer overrun.
In stroke_add_compat() we define ourselves a temporary buffer to
hold points before we 'assign' them properly into the path.
Currently this is set to be 5 points long.

As part of the processing we call line_join_points() to add points
to this. This can add up to 5 points.

Unfortunately, we can call this with 1 point already in the buffer,
meaning that we then overflow by 1 entry.

When we return, we drop the last point - but the overflow has
already happened and the damage has been done.

The fix is simply to extend the temporary buffer by 1 entry.

Credit to Chris for spotting the issue and suggesting the fix.