Bug 707520: Fix detection of Unicode decoding requirement
If the Ghostscript initial device is pdfwrite, the initialisation PostScript
sets up a Unicode decoder to aid in creating (or preserving) a ToUnicode CMap
in the output.

With GhostPDL, the initial device when gs initialises is the nullpage device,
and when we set the "working" device to run a job, we explicitly run
the PS proc ".execute_scheduled_inits" to do those few extra bits of device
specific initialisation.

The problem, in this case, is that we remove the custom operator
".setupUnicodeDecoder" from systemdict at the end of initialisation.

This tweaks that code so it doesn't rely on the operator being present in
systemdict to run.