Tweak icc profile path handling.
We are inconsistent in the way we pass strings and lengths
around for ICC names and lengths. Sometimes we include
the terminator in the length, and sometimes we don't.

This causes problems with Ken's rework of some of the path
handling code, as we now memcpy rather than strcat, and
stray terminators break things. This shows up in the
disable-compile-inits run on the cluster taking an unfeasible
amount of time.

Here we fix various bits of code to consistently NOT include
the terminator. This is consistent with the way we handle
lengthed strings elsewhere in the system.