ps2write - flush cached type 3 glyphs before capturing.
Bug #707554 "PDF to PS segmentation fault without any additional info"

This is a peculiar situation; the problem is caused by executing a glyph
from a type 3 font twice. The first time it is on a page with
transparency, the second it is on a page which does not have
transparency.

The first time we execute it, we are rendering the entire page to a
bitmap, because there is no way to preserve PDF transparency in
PostScript. This means that the ps2write device does NOT see the glyph
description. The interpreter, however, does and caches the glyph.

The second time we execute it, ps2write sets up to capture the glyph
description (to create a font in the output) and runs the glyph.
However, the interpreter knows it has cached the glyph, so it
short-circuits the glyph description and that causes ps2write to fail
because the glyph is not executed.

It turns out we have code designed to deal with situations like thia
already. If we are setting up to capture a glyph we probe the glyph
cache to see if the glyph is cached already. If it is, we flush it from
the cache to make sure the glyph gets executed.

Unfortunately this was intended for a slightly different purpose, and
specifically for the case of the PCL interpreter. See bug #694069
and commit 37b08e30586434098c9f6b8dc35fa25de39f468d

The code as it was works for the original problem, but it actually
relies on some assumptions in PCL which are not true in the general
case. This commit just makes a couple of minor tweaks so that we
properly check the cache and find the glyph.