JPEG handling - remove special case (default) QFactor == 1 handling
Bug #703594 ""QFactor 1" outputs same like "QFactor 0.5""

As per the bug thread, the special handling of QFactor 1 being the
default is incorrect. Although that is the default for the encoder it
can easily be overridden.

Similarly there is code in the decoder which returns the quantisation
tables as a string instead of an array if QFactor is 1. This also makes
no sense to me.

This commit removes the special case handling, if anyone ever turns up
a reason for it then we can revisit it.