PDF interpreter - resolving ColorSpace names, check circular refrences
OSS-fuzz #66711

This is a badly broken file, when we try to resolve a named ColorSpace
(R8) we end up retrieving the exact same name from the Resources
dictionary. This is due to the fact that the Resources ColorSpace
dictionary is itself highly invalid.

We weren't checking that we were getting a different name returned,
which led to us trying to recursively set the named space.

This commit just checks to make sure that if we get a name back from the
Resource lookup, it isn't the same name we were already looking to
resolve.