PDF interpreter - retain Forms with degenerate BBox for pdfwrite
Bug #707579 "PDF file causes Aspose to fail"

I have no clue if this actually addreses the problem in the bug report,
but the output of pdfwrite has a problem with these files.

The problem is that the input files have annotations where the
Appearance stream is a Form XObject with an empty BBox ([0 0 0 0]). An
empty BBox sets up an empty clip, so these Forms will draw nothing. As
an efficiency measure the PDF interpreter simply returns without
running the form in these cases.

Unfortunately this poses a problem for pdfwrite. We set up to capture
the Appearance as a Form XObject in the output PDF file, but then the
Form never gets executed, so we can't reference it, leading to us
outputting an object number of -1, which is illegal.

We can't tell, at the point where we should execute the Form, whether
we are running a Form for an annotation or not, so we can't easily
take different action, and I'd prefer not to plumb that knowledge
through to the Form code.

So in this commit; when executing a form, if the BBox is empty, and we
are preserving high level forms (ie a high level output device, such
as pdfwrite, and forms are being preserved), which we must do for the
annotations, then we execute the begin and end high level form functions
but do not run the form. This will result in an empty form being written
to the output file by pdfwrite.

In addition; when calculating the size and position of an Appearance
stream, if the BBox is empty then discard it and treat it as missing
(which is actually illegal but we have code to cope). This avoid us
doing a divide by zero when attempting to scale the Form.

This resolves the problem I can see with the files in bug 707579.