PDF interpreter - use the correct argument as the properties for DP
No bug report fro this, Chris spotted it while testing soemthing else.

Using annots.pdf, with the pdfwrite device, we were using the 'tag'
as the 'properties' argument to DP. This led to use trying to look up
the tag in the Properties dictionary, but this file doesn't *have* a
Properties dictionary....

Fixed by using the correct argument from the stack.