PostScript interpreter - fix HalftoneMode 1 and 2
Commit 9e2baefb38ba30b5603bb5280e55d45753531096 fixes a problem with
the global VM mode of halftones.

The code to deal with HalftoneMode 1 and 2 is implemented in
PostScript and ends up ignoring the request. However, it evaluates
currenthalftone with the current VM mode now, and then tries to use
that to set the default halftone. If the currnet VM mode isn't global
then that fails because the default halftone is in global VM.

Fix it by ensuring the current VM mode is global. At the same time alter
the 'sethalftone' so that instead of setting the requested halftone and
then throwing it away we just discard it. There's no point in spending
time evaluating a halftone we aren't going to use.

No bug report, uncovered during release testing using the test file
/tests_private/ps/ps3fts/476-01-fixed.ps with the pkmraw device.